Remove unused SVG files and update README
The unused SVG files 'next.svg' and 'vercel.svg' in the 'public' directory have been removed. The README file has been updated with pertinent information about the project including the technologies used, which are Next14, Sanity, and use-shopping-cart. A link to the live site and an image of the project are also included.